Fix UTF-8 page scraping for HTML 5 pages
Unfortunately, DOMDocument is built with a version of libxml that does not understand HTML5 too well. This cause UTF-8 characters to not be read properly. In an attempt to make our scraper work with the majority of HTML5 documents this small change adds a charset and hopefully doesn’t kill efficiency.

I am well aware that this does not 100% solve the problem. If anyone wants to take a stab at this issue I am fine with that given:

1. Some profiling is done to make sure a solution doesn’t make large documents kill our servers.
2. There are no regression bugs. I recommend adding more character cases to the test files or even a non utf-8 test file.